sysutils/opensbi: update to v1.6
Remove the port platform OPTIONS in favor of the 'generic' platform
only. This has been the only one we've shipped with for a few versions
now.

This cleans up the Makefile, but more to the point, there is some kind
of build regression with the new version that doing the make and install
in a single step seems to fix.

Bump dependent u-boot ports.

Sponsored by:	The FreeBSD Foundation