ports-mgmt/porttools: Fix content of distfile
When this port was update to 1.11, value of GH_TAGNAME wasn't
changed.and still points to the commit that is referenced by "1.10"
tag. As a result name of distfile was changed but content was kept
same as that of 1.10. Since defult value of GH_TAGNAME is "1.11" in
this case, fix content of distfile by just removing GH_TAGNAME.

While here,
* Switch to DISTVERSION.
* Pet portlint.
* Tidy up Makefile with portfmt.

PR:		284558
Approved by:	maintiner timeout
Fixes:		2377892144ca ports-mgmt/porttools: update to 1.11