Uses/llvm.mk: remove broken pre-llvm10 support
Once upon a time LLVM releases versions were of the form major.minor as
ports were named things like llvm33 for the 3.3 release.  With LLVM 4,
the project switched to a new major version for each release.  At the
time I kept the form llvm40 for ease of use.  With LLVM 10, I decided it
was time to drop the uniformative minor version to the port bacame
llvm10.  This code attempted to keep LLVM_DEFAULT working for
llvm[789]0 working.  It was broken for llvm10 and now llvm20.  Since
llvm10 was removed last April (6d0c155a81a6) and llvm[789]0 in Jaunary
2023 (21778d8958f5) this code serves no purpose.

Reported by:	Mark Millard <marklmi@yahoo.com>